refactor: move api exceptions into own file
