+++ openedx/core/djangoapps/content_libraries/api/__init__.py
@@ -4,6 +4,7 @@ Python API for working with content libraries
 from .collections import *
 from .containers import *
 from .courseware_import import *
+from .exceptions import *
 from .libraries import *
 from .blocks import *
 from . import permissions
